amadeus2.hr/requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.6
python-slugify==4.0.1
text-unidecode==1.3
Werkzeug==1.0.1
